pyproject.toml
[project]
name = "rapportmea"
version = "0.1.0"
description = ""
readme = "README.md"
requires-python = ">=3.9"
authors = [
    { name = "jk developer", email = "jeantexte20@example.com" }
]
dependencies = [
  "flet==0.28.3",
  "fpdf2",
  "python-docx",
  "flet-permission-handler"
]

[tool.flet]
# org name in reverse domain name notation, e.g. "com.mycompany".
# Combined with project.name to build bundle ID for iOS and Android apps
org = "com.mycompany"

# project display name that is used as an app title on Android and iOS home screens,
# shown in window titles and about app dialogs on desktop.
product = "rapportmea"

# company name to display in about app dialogs
company = "Flet"

# copyright text to display in about app dialogs
copyright = "Copyright (C) 2025 by Flet"

[tool.flet.app]
path = "src"

[tool.uv]
dev-dependencies = [
    "flet[all]==0.28.3",
]

[tool.poetry]
package-mode = false

[tool.flet.android.permission] # --android-permissions
"android.permission.CAMERA" = true
"android.permission.READ_EXTERNAL_STORAGE" = true
"android.WRITE_EXTERNAL_STORAGE" = true
"android.MANAGE_EXTERNAL_STORAGE" = true

[tool.poetry.group.dev.dependencies]
flet = {extras = ["all"], version = "0.28.3"}